% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armadillo_cv_H}
\alias{armadillo_cv_H}
\title{Perform k-fold cross-validation for the bandwidth matrix H}
\usage{
armadillo_cv_H(y, x0, X0, x, X, H_values, k_fold)
}
\arguments{
\item{y}{Output variable}

\item{x0}{Matrix of the predictor variables at a single prediction location}

\item{X0}{Matrix of all the covariates at the prediction location}

\item{x}{Matrix of the predictor variables}

\item{X}{Matrix of all the covariates}

\item{H_values}{List of candidate bandwidth matrices}

\item{k_fold}{Number of folds for cross-validation}
}
\value{
A list containing the selected bandwidth matrix \eqn{\mathbf{H}} and the fitted values using that bandwidth matrix.
}
\description{
This function performs k-fold cross-validation for the bandwidth matrix \eqn{\mathbf{H}} in a local regression model using a Gaussian kernel.
}
