% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armadillo_lm}
\alias{armadillo_lm}
\title{Linear regression using QR decomposition}
\usage{
armadillo_lm(X, y)
}
\arguments{
\item{X}{Model matrix of dimension n x p}

\item{y}{Response vector of length n}
}
\value{
Coefficients of the linear model
}
\description{
Fit a linear regression model with `RcppArmadillo` using QR decomposition.
}
\examples{
data(mtcars)
y <- mtcars$mpg
X <- with(mtcars, cbind(1, wt, wt^2, hp, hp^2))
beta_hat <- armadillo_lm(X, y)
beta_hat
}
