% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armadillo_lm_local}
\alias{armadillo_lm_local}
\title{Local regression model}
\usage{
armadillo_lm_local(y, x0, X0, x, X, H)
}
\arguments{
\item{y}{Output variable}

\item{x0}{Matrix of the predictor variables at a single prediction location}

\item{X0}{Matrix of all the covariates at the prediction location}

\item{x}{Matrix of the predictor variables}

\item{X}{Matrix of all the covariates}

\item{H}{Positive definite bandwidth matrix}
}
\value{
The estimated regression coefficients \eqn{\hat \beta(\mathbf{x}_0)} at the prediction locations
}
\description{
This function is used to fit a local regression model with Gaussian kernel weights using `RcppArmadillo`.
For a given \eqn{\mathbf{x}_0}, this function estimates \eqn{\hat{\beta}(\mathbf{x}_0)}, such that
\deqn{\hat{\beta}(\mathbf{x}_0) = \arg\min_{\beta} \sum_{i=1}^n \kappa_{\mathbf{H}}(\mathbf{x}_0 - \mathbf{x}_i) (y_i - \tilde{\mathbf{x}}_i^\top \beta)^2}
where \eqn{\kappa_H} is a density kernel with positive definite bandwidth matrix \eqn{\mathbf{H}}.
}
\examples{
data(mtcars)
y <- mtcars$mpg
x <- as.matrix(mtcars[, c("wt", "hp")])
X <- with(mtcars, cbind(1, wt, wt^2, hp, hp^2))
H <- diag(2) # use identity bandwidth matrix for example
# Predict at three locations
preds <- armadillo_lm_local(y= y, x0 = x[1:3,], X0 = X[1:3,], x = x, X = X, H = H)
}
