% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadilLocalregression-package.R
\docType{package}
\name{armadilLocalregression-package}
\alias{armadilLocalregression}
\alias{armadilLocalregression-package}
\title{armadilLocalregression: Local Polynomial Regression Using RcppArmadillo}
\description{
Functions for linear regression and local polynomial regression using RcppArmadillo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/babichmorrowc/armadilLocalregression}
  \item Report bugs at \url{https://github.com/babichmorrowc/armadilLocalregression/issues}
}

}
\keyword{internal}
