% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmLocal.R
\name{lmLocal}
\alias{lmLocal}
\title{Local least squares regression with Gaussian kernel}
\usage{
lmLocal(y, x0, X0, x, X, H)
}
\arguments{
\item{y}{Output variable}

\item{x0}{Matrix of the predictor variables at the prediction locations}

\item{X0}{Matrix of all the covariates at the prediction locations}

\item{x}{Matrix of the predictor variables}

\item{X}{Matrix of all the covariates}

\item{H}{Positive definite bandwidth matrix}
}
\value{
The estimated regression coefficients \eqn{\hat \beta(\mathbf{x}_0)} at the prediction locations
}
\description{
This function fits a local linear regression model with Gaussian kernel weights.
The regression coefficients are a function of the covariates \eqn{\mathbf{x}}, i.e. \eqn{\hat{\beta} = \hat{\beta}(\mathbf{x})}.
For a given \eqn{\mathbf{x}_0}, \deqn{\hat{\beta}(\mathbf{x}_0) = \arg\min_{\beta} \sum_{i=1}^n \kappa_{\mathbf{H}}(\mathbf{x}_0 - \mathbf{x}_i) (y_i - \tilde{\mathbf{x}}_i^\top \beta)^2}
where \eqn{\kappa_H} is a density kernel with positive definite bandwidth matrix \eqn{\mathbf{H}}.
}
